+++ src/components/Select/Select.tsx
@@ -1,12 +1,11 @@
 import React from 'react';
 
 import {KeyCode} from '../../constants';
-import {useForkRef, useUniqId} from '../../hooks';
+import {useForkRef, useSelect, useUniqId} from '../../hooks';
 import type {List} from '../List';
 import {useMobile} from '../mobile';
 import type {CnMods} from '../utils/cn';
 import {useFocusWithin} from '../utils/interactions';
-import {useSelect} from '../utils/useSelect';
 
 import {EmptyOptions, SelectControl, SelectFilter, SelectList, SelectPopup} from './components';
 import {DEFAULT_VIRTUALIZATION_THRESHOLD, selectBlock} from './constants';