+++ src/components/utils/useListNavigation/index.ts
@@ -1,1 +1,0 @@
-export {useListNavigation} from './useListNavigation';