+++ src/components/utils/useViewportSize/index.ts
@@ -1,2 +1,0 @@
-export {useViewportSize} from './useViewportSize';
-export type {ViewportSize} from './useViewportSize';