+++ src/demo/Showcase/Showcase.tsx
@@ -5,7 +5,7 @@ import {cn} from '../../components/utils/cn';
 import './Showcase.scss';
 
 type Props = React.PropsWithChildren<{
-    title: string;
+    title?: string;
     description?: React.ReactNode;
     className?: string;
 }>;
@@ -15,8 +15,8 @@ const b = cn('showcase');
 export function Showcase({title, description, className, children}: Props) {
     return (
         <div className={b(null, className)}>
-            <div className={b('title')}>{title}</div>
-            <div className={b('description')}>{description}</div>
+            {title && <div className={b('title')}>{title}</div>}
+            {description && <div className={b('description')}>{description}</div>}
             <div className={b('content')}>{children}</div>
         </div>
     );