+++ src/hooks/useIntersection/index.ts
@@ -1,0 +1,2 @@
+export {useIntersection} from './useIntersection';
+export type {UseIntersection, UseIntersectionProps} from './useIntersection';