+++ src/hooks/useViewportSize/index.ts
@@ -1,0 +1,2 @@
+export {useViewportSize} from './useViewportSize';
+export type {UseViewportSizeResult} from './useViewportSize';